// src/interfaces/Icomic.ts
export interface ComicInterface {
    month:      string;
    num:        number;
    link:       string;
    year:       string;
    news:       string;
    safe_title: string;
    transcript: string;
    alt:        string;
    img:        string;
    title:      string;
    day:        string;
}
  